+++ frontend/src/app/features/main-page/components/description-section/components/game/game.component.ts
@@ -1,7 +1,12 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
+
+import { catchError, EMPTY, take } from 'rxjs';
 
 import { GameCategory } from '../../../../../../interfaces/game-category.interface';
 import { EditGameService } from '../../../../services/edit-game.service';
+import { SnackbarService } from './../../../../../../components/snackbar/snackbar.service';
+import { DeleteGameRequestBody } from './interfaces/delete-game-request-body.interface';
 
 @Component({
     selector: 'mados-game',
@@ -19,9 +24,33 @@ export class GameComponent {
     @Input()
     public shouldShowEdition: boolean;
 
-    constructor(private editGameService: EditGameService) {}
+    @Output()
+    public reloadDescription: EventEmitter<void> = new EventEmitter<void>();
+
+    constructor(
+        private editGameService: EditGameService,
+        private httpClient: HttpClient,
+        private snackbarService: SnackbarService,
+    ) {}
 
     public editGame(): void {
         this.editGameService.editingGameName$.next(this.name);
+    }
+
+    public deleteGame(): void {
+        const requestBody: DeleteGameRequestBody = { name: this.name };
+
+        this.httpClient.delete('game', { body: requestBody })
+            .pipe(
+                take(1),
+                catchError(() => {
+                    this.snackbarService.openSnackbar('An error has occurred while deleting game', 'error');
+
+                    return EMPTY;
+                })
+            ).subscribe(() => {
+                this.snackbarService.openSnackbar('Successfully deleted game', 'success');
+                this.reloadDescription.emit();
+            });
     }
 }