+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Both frontend and backend changes and the final version of the application will be documented in this file.
 
+## 0.21.0 - 20.02.2022 [FE]
+
+- Add option to delete specific game
+- Add infobar when there is no data
+- Change 'add-game' endpoint name to 'game'
+- Fix setting initial categories when there is no data
+- Small refactor
+
 ## 0.20.0 - 20.02.2022 [FE]
 
 - Add snackbars